Moved variables functions to new file
Moved the functions related to variables to a new file called `vars.py`. Then fixed calls to the functions and cleaned up docstrings, and removed the old functions from the readers.
